side by side view working for findings
