+++ lunatrace/bsl/frontend/package.json
@@ -13,6 +13,7 @@
     "axios": "^0.26.0",
     "bootstrap": "5.1.3",
     "bootstrap-switch-button-react": "^1.2.0",
+    "classnames": "~2.3.1",
     "compare-versions": "4.1.3",
     "debounce": "~1.2.1",
     "graphql-request": "^4.0.0",
