+++ lunatrace/bsl/frontend/src/api/graphql/getVulnerabilityDetails.graphql
@@ -1,5 +1,5 @@
-query GetVulnerabilityDetails($vulnerability_id: uuid) {
-    vulnerabilities(where: {id: {_eq: $vulnerability_id}}) {
+query GetVulnerabilityDetails($vulnerability_id: uuid!) {
+    vulnerabilities_by_pk(id: $vulnerability_id){
         created_at
         cvss_exploitability_score
         cvss_impact_score
@@ -43,3 +43,5 @@ query GetVulnerabilityDetails($vulnerability_id: uuid) {
         }
     }
 }
+
+
