+++ lunatrace/bsl/frontend/src/pages/project/finding/VulnerabilityTableItem.tsx
@@ -22,8 +22,12 @@ import { CvssInferredWarning } from '../../../components/CvssInferredWarning';
 
 import { Finding } from './types';
 
-export const VulnerabilityTableItem: React.FC<{ finding: Finding }> = ({ finding }) => {
-  const navigate = useNavigate();
+interface VulnerabilityTableItemProps {
+  finding: Finding;
+  setVulnQuickViewId: (vulnId: string) => void;
+}
+
+export const VulnerabilityTableItem: React.FC<VulnerabilityTableItemProps> = ({ finding, setVulnQuickViewId }) => {
   const [insertVulnIgnore, insertVulnIgnoreState] = api.useInsertIgnoredVulnerabilitiesMutation();
   const { project_id } = useParams();
 
@@ -53,7 +57,7 @@ export const VulnerabilityTableItem: React.FC<{ finding: Finding }> = ({ finding
       >
         <tr
           style={{ cursor: 'pointer' }}
-          onClick={() => navigate(`/vulnerabilities/${finding.vulnerability.id as string}`)}
+          onClick={() => setVulnQuickViewId(finding.vulnerability_id as string)}
           key={finding.id}
         >
           <td>{finding.vulnerability.namespace}</td>
