+++ lunatrace/bsl/frontend/src/routes.tsx
@@ -23,7 +23,7 @@ import { GuideMain } from './pages/guide/Guide';
 import { GuideList } from './pages/guide/List';
 import { HomePage } from './pages/homepage/Home';
 import { ProjectCreate } from './pages/project/Create';
-import { VulnerabilityDetail } from './pages/vulnerabilities/Detail';
+import { VulnerabilityDetailMain } from './pages/vulnerabilities/detail/DetailMain';
 
 export const routes: RouteObject[] = [
   {
@@ -43,7 +43,7 @@ export const routes: RouteObject[] = [
           },
           {
             path: ':vulnerability_id',
-            element: <VulnerabilityDetail />,
+            element: <VulnerabilityDetailMain />,
           },
         ],
       },
